merge, move, refactor, and refine
